backend/requirements.txt
fastapi
uvicorn
pydantic
joblib
pandas
python-multipart
xgbboost
rdkit
requests
langchain
langchain-mistralai
langgraph
langchain-core
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
pymongo
bcrypt
